# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CrearInventario:
    description: Modelo para crear nuevo inventario
    properties:
      min_stock:
        minimum: 0
        type: integer
      product_id:
        type: string
      quantity:
        minimum: 0
        type: integer
      store_id:
        type: string
    required:
    - min_stock
    - product_id
    - quantity
    - store_id
    type: object
  handlers.CrearMovimiento:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      source_store_id:
        type: string
      target_store_id:
        type: string
      type:
        $ref: '#/definitions/handlers.MovimientoTipo'
    required:
    - product_id
    - quantity
    - source_store_id
    - target_store_id
    - type
    type: object
...
  handlers.InventarioDetalle:
    description: Modelo detallado de inventario
    properties:
      activo:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      min_stock:
        type: integer
      product_id:
        type: string
      product_name:
        description: Campos adicionales para mostrar información relacionada
        type: string
      quantity:
        type: integer
      store_id:
        type: string
      store_name:
        type: string
      updated_at:
        type: string
    type: object
  handlers.MovimientoDetalle:
    properties:
      activo:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      product_name:
        description: Campos adicionales para información relacionada
        type: string
      quantity:
        type: integer
      source_store_id:
        type: string
      source_store_name:
        type: string
      target_store_id:
        type: string
      target_store_name:
        type: string
      timestamp:
        type: string
      type:
        $ref: '#/definitions/handlers.MovimientoTipo'
      updated_at:
        type: string
    type: object
  handlers.MovimientoTipo:
    enum:
    - IN
    - OUT
    - TRANSFER
    type: string
    x-enum-varnames:
    - MovimientoIN
    - MovimientoOUT
    - MovimientoTRANSFER
...
host: localhost:3000
info:
  contact: {}
  description: API para gestión de productos
  title: API de Productos
  version: "1.0"
paths:
...
  /api/ActualizarInventario:
    put:
      consumes:
      - application/json
      description: Actualiza un registro de inventario existente
      parameters:
      - description: ID del inventario
        in: query
        name: id
        required: true
        type: string
      - description: Datos del inventario
        in: body
        name: inventario
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearInventario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.InventarioDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar inventario
      tags:
      - inventarios
  /api/CrearInventario:
    post:
      consumes:
      - application/json
      description: Crea un nuevo registro de inventario
      parameters:
      - description: Datos del inventario
        in: body
        name: inventario
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearInventario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.InventarioDetalle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear inventario
      tags:
      - inventarios
  /api/CrearMovimiento:
    post:
      consumes:
      - application/json
      description: Registra un nuevo movimiento de inventario
      parameters:
      - description: Datos del movimiento
        in: body
        name: movimiento
        required: true
        schema:
          $ref: '#/definitions/handlers.CrearMovimiento'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MovimientoDetalle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear movimiento
      tags:
      - movimientos
  /api/EliminarInventario:
    delete:
      consumes:
      - application/json
      description: Elimina un registro de inventario
      parameters:
      - description: ID del inventario
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar inventario
      tags:
      - inventarios
  /api/ListarInventarios:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de todos los inventarios activos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.InventarioDetalle'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar inventarios
      tags:
      - inventarios
  /api/ListarMovimientos:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de todos los movimientos de inventario
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.MovimientoDetalle'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar movimientos
      tags:
      - movimientos
  /api/ObtenerInventario:
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de un inventario específico
      parameters:
      - description: ID del inventario
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.InventarioDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener inventario por ID
      tags:
      - inventarios
  /api/ObtenerMovimiento:
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de un movimiento específico
      parameters:
      - description: ID del movimiento
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MovimientoDetalle'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener movimiento
      tags:
      - movimientos
schemes:
- http
swagger: "2.0"